# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stories API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/stories/{story_id}/comments/{comment_id}/reactions": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получение списка реакций к определенному комментарию",
                "operationId": "3dc22ec30bcd7aefbfb759f831d4a91e",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Номер телефона",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "7000801"
                    },
                    {
                        "name": "story_id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "31"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 13
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "7000801"
                                                    },
                                                    "story_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "comment_id": {
                                                        "type": "integer",
                                                        "example": 31
                                                    },
                                                    "emoji": {
                                                        "type": "string",
                                                        "example": "😯"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Создание реакции к определенному комментарию",
                "operationId": "2a0f18f90760dc9e81b2a00e5226cd70",
                "parameters": [
                    {
                        "name": "story_id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    },
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "31"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "phone": {
                                                "type": "string",
                                                "example": "7000801"
                                            },
                                            "emoji": {
                                                "type": "string",
                                                "example": "🙌"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "event": {
                                            "type": "string",
                                            "example": "CREATED"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Реакция 🙌 была успешна добавлена к комментарию"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 19
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "7000803"
                                                },
                                                "story_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "comment_id": {
                                                    "type": "integer",
                                                    "example": 31
                                                },
                                                "emoji": {
                                                    "type": "string",
                                                    "example": "🙌"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Удаление комментария по ID у определенного сториса",
                "operationId": "91c3c34ece9bd0afbd8ac042512ec5e9",
                "parameters": [
                    {
                        "name": "story_id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    },
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "31"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "phone": {
                                                "type": "string",
                                                "example": "7000801"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "DELETED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "event": {
                                            "type": "string",
                                            "example": "DELETED"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Реакция была успешно удалена"
                                        },
                                        "data": {
                                            "properties": {
                                                "phone": {
                                                    "type": "string",
                                                    "example": "7000803"
                                                },
                                                "story_id": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "comment_id": {
                                                    "type": "integer",
                                                    "example": "31"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "apiKey",
                "in": "header"
            }
        }
    },
...
}